need to clean up css in booking form and recreate available slot after appointment deletion

+++ client/src/components/Book/CancelBooking/CancelBooking.js
@@ -12,10 +12,22 @@ const CancelBooking = () => {
     setStatus('loading');
     let confirmationNumber = form['Confirmation #'];
     try {
-      const res = await axios.delete(
-        `/api/booking/${confirmationNumber}`
+      const findSlot = await axios.get(
+        `/api/booking/find/${confirmationNumber}`
       );
-      console.log('deleted booking', res.data)
+      console.log('findSlot', findSlot)
+
+      // const makeAvailable = findSlot.data.availability;
+      // console.log('makeAvailable', makeAvailable)
+
+      // const updateSlot = await axios.put(
+      //   `/api/availability/true/${findSlot.data.availability}`
+      // );
+      // console.log('updated slot', updateSlot.data);
+
+      const res = await axios.delete(`/api/booking/${confirmationNumber}`);
+      console.log('deleted booking', res.data);
+
       await setTimeout(() => {
         setStatus('success');
       }, 3000);
